students/K3339/Skirlyak_Yaroslav/lab1/requirements.txt
fastapi==0.115.0
sqlmodel==0.0.22
psycopg2-binary==2.9.10
alembic==1.13.3
passlib[bcrypt]==1.7.4
python-jose==3.3.0
pycryptodome==3.20.0
python-dotenv==1.0.1
uvicorn==0.30.6